// src/api/nvd/cves/2.0/schema/response.ts
export type CveId = string;
export type JSONSchemaForCommonVulnerabilityScoringSystemVersion40 = (
  | {
      baseScore?: number;
      baseSeverity?: "NONE";
      [k: string]: unknown;
    }
  | {
      baseScore?: number;
      baseSeverity?: "LOW";
      [k: string]: unknown;
    }
  | {
      baseScore?: number;
      baseSeverity?: "MEDIUM";
      [k: string]: unknown;
    }
  | {
      baseScore?: number;
      baseSeverity?: "HIGH";
      [k: string]: unknown;
    }
  | {
      baseScore?: number;
      baseSeverity?: "CRITICAL";
      [k: string]: unknown;
    }
) &
  (
    | {
        threatScore?: number;
        threatSeverity?: "NONE";
        [k: string]: unknown;
      }
    | {
        threatScore?: number;
        threatSeverity?: "LOW";
        [k: string]: unknown;
      }
    | {
        threatScore?: number;
        threatSeverity?: "MEDIUM";
        [k: string]: unknown;
      }
    | {
        threatScore?: number;
        threatSeverity?: "HIGH";
        [k: string]: unknown;
      }
    | {
        threatScore?: number;
        threatSeverity?: "CRITICAL";
        [k: string]: unknown;
      }
  ) &
  (
    | {
        environmentalScore?: number;
        environmentalSeverity?: "NONE";
        [k: string]: unknown;
      }
    | {
        environmentalScore?: number;
        environmentalSeverity?: "LOW";
        [k: string]: unknown;
      }
    | {
        environmentalScore?: number;
        environmentalSeverity?: "MEDIUM";
        [k: string]: unknown;
      }
    | {
        environmentalScore?: number;
        environmentalSeverity?: "HIGH";
        [k: string]: unknown;
      }
    | {
        environmentalScore?: number;
        environmentalSeverity?: "CRITICAL";
        [k: string]: unknown;
      }
  ) & {
    /**
     * CVSS Version
     */
    version: "4.0";
    vectorString: string;
    attackVector?: "NETWORK" | "ADJACENT" | "LOCAL" | "PHYSICAL";
    attackComplexity?: "HIGH" | "LOW";
    attackRequirements?: "NONE" | "PRESENT";
    privilegesRequired?: "HIGH" | "LOW" | "NONE";
    userInteraction?: "NONE" | "PASSIVE" | "ACTIVE";
    vulnConfidentialityImpact?: "NONE" | "LOW" | "HIGH";
    vulnIntegrityImpact?: "NONE" | "LOW" | "HIGH";
    vulnAvailabilityImpact?: "NONE" | "LOW" | "HIGH";
    subConfidentialityImpact?: "NONE" | "LOW" | "HIGH";
    subIntegrityImpact?: "NONE" | "LOW" | "HIGH";
    subAvailabilityImpact?: "NONE" | "LOW" | "HIGH";
    exploitMaturity?: "UNREPORTED" | "PROOF_OF_CONCEPT" | "ATTACKED" | "NOT_DEFINED";
    confidentialityRequirement?: "LOW" | "MEDIUM" | "HIGH" | "NOT_DEFINED";
    integrityRequirement?: "LOW" | "MEDIUM" | "HIGH" | "NOT_DEFINED";
    availabilityRequirement?: "LOW" | "MEDIUM" | "HIGH" | "NOT_DEFINED";
    modifiedAttackVector?: "NETWORK" | "ADJACENT" | "LOCAL" | "PHYSICAL" | "NOT_DEFINED";
    modifiedAttackComplexity?: "HIGH" | "LOW" | "NOT_DEFINED";
    modifiedAttackRequirements?: "NONE" | "PRESENT" | "NOT_DEFINED";
    modifiedPrivilegesRequired?: "HIGH" | "LOW" | "NONE" | "NOT_DEFINED";
    modifiedUserInteraction?: "NONE" | "PASSIVE" | "ACTIVE" | "NOT_DEFINED";
    modifiedVulnConfidentialityImpact?: "NONE" | "LOW" | "HIGH" | "NOT_DEFINED";
    modifiedVulnIntegrityImpact?: "NONE" | "LOW" | "HIGH" | "NOT_DEFINED";
    modifiedVulnAvailabilityImpact?: "NONE" | "LOW" | "HIGH" | "NOT_DEFINED";
    modifiedSubConfidentialityImpact?: "NEGLIGIBLE" | "LOW" | "HIGH" | "NOT_DEFINED";
    modifiedSubIntegrityImpact?: "NEGLIGIBLE" | "LOW" | "HIGH" | "SAFETY" | "NOT_DEFINED";
    modifiedSubAvailabilityImpact?: "NEGLIGIBLE" | "LOW" | "HIGH" | "SAFETY" | "NOT_DEFINED";
    Safety?: "NEGLIGIBLE" | "PRESENT" | "NOT_DEFINED";
    Automatable?: "NO" | "YES" | "NOT_DEFINED";
    Recovery?: "AUTOMATIC" | "USER" | "IRRECOVERABLE" | "NOT_DEFINED";
    valueDensity?: "DIFFUSE" | "CONCENTRATED" | "NOT_DEFINED";
    vulnerabilityResponseEffort?: "LOW" | "MODERATE" | "HIGH" | "NOT_DEFINED";
    providerUrgency?: "CLEAR" | "GREEN" | "AMBER" | "RED" | "NOT_DEFINED";
    [k: string]: unknown;
  };
/**
 * CVSS subscore.
 */
export type DefSubscore = number;

export interface JSONSchemaForNVDVulnerabilityDataAPIVersion223 {
  resultsPerPage: number;
  startIndex: number;
  totalResults: number;
  format: string;
  version: string;
  timestamp: string;
  /**
   * NVD feed array of CVE
   */
  vulnerabilities: DefCveItem[];
}
export interface DefCveItem {
  cve: CveItem;
}
export interface CveItem {
  id: CveId;
  sourceIdentifier?: string;
  vulnStatus?: string;
  published: string;
  lastModified: string;
  evaluatorComment?: string;
  evaluatorSolution?: string;
  evaluatorImpact?: string;
  cisaExploitAdd?: string;
  cisaActionDue?: string;
  cisaRequiredAction?: string;
  cisaVulnerabilityName?: string;
  cveTags?: {
    /**
     * The email address or UUID of the source that contributed the information
     */
    sourceIdentifier?: string;
    tags?: ("unsupported-when-assigned" | "exclusively-hosted-service" | "disputed")[];
    [k: string]: unknown;
  }[];
  /**
   * @minItems 1
   */
  descriptions: [LangString, ...LangString[]];
  references: Reference[];
  /**
   * Metric scores for a vulnerability as found on NVD.
   */
  metrics?: {
    /**
     * CVSS V4.0 score.
     */
    cvssMetricV40?: CvssV40[];
    /**
     * CVSS V3.1 score.
     */
    cvssMetricV31?: CvssV31[];
    /**
     * CVSS V3.0 score.
     */
    cvssMetricV30?: CvssV30[];
    /**
     * CVSS V2.0 score.
     */
    cvssMetricV2?: CvssV2[];
    [k: string]: unknown;
  };
  weaknesses?: Weakness[];
  configurations?: Config[];
  vendorComments?: VendorComment[];
  [k: string]: unknown;
}
export interface LangString {
  lang: string;
  value: string;
}
export interface Reference {
  url: string;
  source?: string;
  tags?: string[];
  [k: string]: unknown;
}
export interface CvssV40 {
  source: string;
  type: "Primary" | "Secondary";
  cvssData: JSONSchemaForCommonVulnerabilityScoringSystemVersion40;
}
export interface CvssV31 {
  source: string;
  type: "Primary" | "Secondary";
  cvssData: JSONSchemaForCommonVulnerabilityScoringSystemVersion31;
  exploitabilityScore?: DefSubscore;
  impactScore?: DefSubscore;
}
export interface JSONSchemaForCommonVulnerabilityScoringSystemVersion31 {
  /**
   * CVSS Version
   */
  version: "3.1";
  vectorString: string;
  attackVector?: "NETWORK" | "ADJACENT_NETWORK" | "LOCAL" | "PHYSICAL";
  attackComplexity?: "HIGH" | "LOW";
  privilegesRequired?: "HIGH" | "LOW" | "NONE";
  userInteraction?: "NONE" | "REQUIRED";
  scope?: "UNCHANGED" | "CHANGED";
  confidentialityImpact?: "NONE" | "LOW" | "HIGH";
  integrityImpact?: "NONE" | "LOW" | "HIGH";
  availabilityImpact?: "NONE" | "LOW" | "HIGH";
  baseScore: number;
  baseSeverity: "NONE" | "LOW" | "MEDIUM" | "HIGH" | "CRITICAL";
  exploitCodeMaturity?: "UNPROVEN" | "PROOF_OF_CONCEPT" | "FUNCTIONAL" | "HIGH" | "NOT_DEFINED";
  remediationLevel?: "OFFICIAL_FIX" | "TEMPORARY_FIX" | "WORKAROUND" | "UNAVAILABLE" | "NOT_DEFINED";
  reportConfidence?: "UNKNOWN" | "REASONABLE" | "CONFIRMED" | "NOT_DEFINED";
  temporalScore?: number;
  temporalSeverity?: "NONE" | "LOW" | "MEDIUM" | "HIGH" | "CRITICAL";
  confidentialityRequirement?: "LOW" | "MEDIUM" | "HIGH" | "NOT_DEFINED";
  integrityRequirement?: "LOW" | "MEDIUM" | "HIGH" | "NOT_DEFINED";
  availabilityRequirement?: "LOW" | "MEDIUM" | "HIGH" | "NOT_DEFINED";
  modifiedAttackVector?: "NETWORK" | "ADJACENT_NETWORK" | "LOCAL" | "PHYSICAL" | "NOT_DEFINED";
  modifiedAttackComplexity?: "HIGH" | "LOW" | "NOT_DEFINED";
  modifiedPrivilegesRequired?: "HIGH" | "LOW" | "NONE" | "NOT_DEFINED";
  modifiedUserInteraction?: "NONE" | "REQUIRED" | "NOT_DEFINED";
  modifiedScope?: "UNCHANGED" | "CHANGED" | "NOT_DEFINED";
  modifiedConfidentialityImpact?: "NONE" | "LOW" | "HIGH" | "NOT_DEFINED";
  modifiedIntegrityImpact?: "NONE" | "LOW" | "HIGH" | "NOT_DEFINED";
  modifiedAvailabilityImpact?: "NONE" | "LOW" | "HIGH" | "NOT_DEFINED";
  environmentalScore?: number;
  environmentalSeverity?: "NONE" | "LOW" | "MEDIUM" | "HIGH" | "CRITICAL";
  [k: string]: unknown;
}
export interface CvssV30 {
  source: string;
  type: "Primary" | "Secondary";
  cvssData: JSONSchemaForCommonVulnerabilityScoringSystemVersion30;
  exploitabilityScore?: DefSubscore;
  impactScore?: DefSubscore;
}
export interface JSONSchemaForCommonVulnerabilityScoringSystemVersion30 {
  /**
   * CVSS Version
   */
  version: "3.0";
  vectorString: string;
  attackVector?: "NETWORK" | "ADJACENT_NETWORK" | "LOCAL" | "PHYSICAL";
  attackComplexity?: "HIGH" | "LOW";
  privilegesRequired?: "HIGH" | "LOW" | "NONE";
  userInteraction?: "NONE" | "REQUIRED";
  scope?: "UNCHANGED" | "CHANGED";
  confidentialityImpact?: "NONE" | "LOW" | "HIGH";
  integrityImpact?: "NONE" | "LOW" | "HIGH";
  availabilityImpact?: "NONE" | "LOW" | "HIGH";
  baseScore: number;
  baseSeverity: "NONE" | "LOW" | "MEDIUM" | "HIGH" | "CRITICAL";
  exploitCodeMaturity?: "UNPROVEN" | "PROOF_OF_CONCEPT" | "FUNCTIONAL" | "HIGH" | "NOT_DEFINED";
  remediationLevel?: "OFFICIAL_FIX" | "TEMPORARY_FIX" | "WORKAROUND" | "UNAVAILABLE" | "NOT_DEFINED";
  reportConfidence?: "UNKNOWN" | "REASONABLE" | "CONFIRMED" | "NOT_DEFINED";
  temporalScore?: number;
  temporalSeverity?: "NONE" | "LOW" | "MEDIUM" | "HIGH" | "CRITICAL";
  confidentialityRequirement?: "LOW" | "MEDIUM" | "HIGH" | "NOT_DEFINED";
  integrityRequirement?: "LOW" | "MEDIUM" | "HIGH" | "NOT_DEFINED";
  availabilityRequirement?: "LOW" | "MEDIUM" | "HIGH" | "NOT_DEFINED";
  modifiedAttackVector?: "NETWORK" | "ADJACENT_NETWORK" | "LOCAL" | "PHYSICAL" | "NOT_DEFINED";
  modifiedAttackComplexity?: "HIGH" | "LOW" | "NOT_DEFINED";
  modifiedPrivilegesRequired?: "HIGH" | "LOW" | "NONE" | "NOT_DEFINED";
  modifiedUserInteraction?: "NONE" | "REQUIRED" | "NOT_DEFINED";
  modifiedScope?: "UNCHANGED" | "CHANGED" | "NOT_DEFINED";
  modifiedConfidentialityImpact?: "NONE" | "LOW" | "HIGH" | "NOT_DEFINED";
  modifiedIntegrityImpact?: "NONE" | "LOW" | "HIGH" | "NOT_DEFINED";
  modifiedAvailabilityImpact?: "NONE" | "LOW" | "HIGH" | "NOT_DEFINED";
  environmentalScore?: number;
  environmentalSeverity?: "NONE" | "LOW" | "MEDIUM" | "HIGH" | "CRITICAL";
  [k: string]: unknown;
}
export interface CvssV2 {
  source: string;
  type: "Primary" | "Secondary";
  cvssData: JSONSchemaForCommonVulnerabilityScoringSystemVersion20;
  baseSeverity?: string;
  exploitabilityScore?: DefSubscore;
  impactScore?: DefSubscore;
  acInsufInfo?: boolean;
  obtainAllPrivilege?: boolean;
  obtainUserPrivilege?: boolean;
  obtainOtherPrivilege?: boolean;
  userInteractionRequired?: boolean;
}
export interface JSONSchemaForCommonVulnerabilityScoringSystemVersion20 {
  /**
   * CVSS Version
   */
  version: "2.0";
  vectorString: string;
  accessVector?: "NETWORK" | "ADJACENT_NETWORK" | "LOCAL";
  accessComplexity?: "HIGH" | "MEDIUM" | "LOW";
  authentication?: "MULTIPLE" | "SINGLE" | "NONE";
  confidentialityImpact?: "NONE" | "PARTIAL" | "COMPLETE";
  integrityImpact?: "NONE" | "PARTIAL" | "COMPLETE";
  availabilityImpact?: "NONE" | "PARTIAL" | "COMPLETE";
  baseScore: number;
  exploitability?: "UNPROVEN" | "PROOF_OF_CONCEPT" | "FUNCTIONAL" | "HIGH" | "NOT_DEFINED";
  remediationLevel?: "OFFICIAL_FIX" | "TEMPORARY_FIX" | "WORKAROUND" | "UNAVAILABLE" | "NOT_DEFINED";
  reportConfidence?: "UNCONFIRMED" | "UNCORROBORATED" | "CONFIRMED" | "NOT_DEFINED";
  temporalScore?: number;
  collateralDamagePotential?: "NONE" | "LOW" | "LOW_MEDIUM" | "MEDIUM_HIGH" | "HIGH" | "NOT_DEFINED";
  targetDistribution?: "NONE" | "LOW" | "MEDIUM" | "HIGH" | "NOT_DEFINED";
  confidentialityRequirement?: "LOW" | "MEDIUM" | "HIGH" | "NOT_DEFINED";
  integrityRequirement?: "LOW" | "MEDIUM" | "HIGH" | "NOT_DEFINED";
  availabilityRequirement?: "LOW" | "MEDIUM" | "HIGH" | "NOT_DEFINED";
  environmentalScore?: number;
  [k: string]: unknown;
}
export interface Weakness {
  source: string;
  type: string;
  /**
   * @minItems 0
   */
  description: LangString[];
}
export interface Config {
  operator?: "AND" | "OR";
  negate?: boolean;
  nodes: Node[];
  [k: string]: unknown;
}
/**
 * Defines a configuration node in an NVD applicability statement.
 */
export interface Node {
  operator: "AND" | "OR";
  negate?: boolean;
  cpeMatch: CpeMatch[];
  [k: string]: unknown;
}
/**
 * CPE match string or range
 */
export interface CpeMatch {
  vulnerable: boolean;
  criteria: string;
  matchCriteriaId: string;
  versionStartExcluding?: string;
  versionStartIncluding?: string;
  versionEndExcluding?: string;
  versionEndIncluding?: string;
  [k: string]: unknown;
}
export interface VendorComment {
  organization: string;
  comment: string;
  lastModified: string;
}